backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
paho-mqtt==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.7.0
pydantic-settings==2.2.0
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
